version 1.0.0

+++ PeppyMeter/vumeter.py
@@ -225,8 +225,8 @@ class Vumeter(ScreensaverMeter):
     def refresh(self):
         """ Refresh meter. Used to update random meter. """
         if(self.frames%self.util.meter_config[FRAME_RATE]==0):
-
-            os.kill(self.audiotopPID, signal.SIGALRM)
+            if not self.testMode:
+                os.kill(self.audiotopPID, signal.SIGALRM)
 
             self.sio.emit('getState', {})
             if self.playerstatus == STARTPLAYING or self.playerstatus == PLAYING:
